Divides astarStructs into 2 files and adds an InversedComparator for Node
