+++ Makefile
@@ -1,5 +1,5 @@
 CXXFLAGS = -Wall -Werror -Wextra -g -std=c++20
-OBJECTS = obj/astar.o obj/astarStructs.o
+OBJECTS = obj/astar.o obj/astarStructs.o obj/astarNodes.o
 
 .PHONY: all clean
 
@@ -13,7 +13,8 @@ $(OBJECTS): obj/%.o: src/%.cpp
 	g++ $(CXXFLAGS) -I include -c $< -o $@
 
 astar.o : astarStructs.h
-astarStructs.o : astarStructs.h
+astarStructs.o : astarStructs.h astarNodes.h
+astarNodes.o : astarNodes.h
 
 algo : obj $(OBJECTS)
 	g++ $(CXXFLAGS) -o $@ $(OBJECTS)
