+++ include/astarStructs.h
@@ -2,7 +2,6 @@
 
 #include <algorithm>
 #include <cmath>
-#include <vector>
 #include <exception>
 #include <functional>
 #include <iostream>
@@ -10,78 +9,13 @@
 #include <string>
 #include <tuple>
 #include <unordered_map>
-#include <unordered_set>
 #include <utility>
 #include <vector>
 
+#include "astarNodes.h"
+
 
 int manhattan(int i1, int j1, int i2, int j2);
-
-
-struct BaseNode {
-    friend bool operator==(BaseNode const& first, BaseNode const& second);
-    friend bool operator<(BaseNode const& first, BaseNode const& second);
-    friend std::ostream& operator<<(std::ostream& os, BaseNode const& baseNode);
-public:
-    BaseNode(int i = 0, int j = 0);
-    std::tuple<int, int> getTuple() const;
-
-public:
-    int _i;
-    int _j;
-};
-
-bool operator==(BaseNode const& first, BaseNode const& second);
-bool operator<(BaseNode const& first, BaseNode const& second);
-std::ostream& operator<<(std::ostream& os, BaseNode const& node);
-
-struct hashBaseNode {
-    typedef BaseNode argument_type;
-    typedef std::size_t result_type;
-
-    result_type operator()(argument_type const& s) const;
-};
-
-
-class Node {
-    friend bool operator==(Node const& first, Node const& second);
-    friend bool operator<(Node const& first, Node const& second);
-    friend std::ostream& operator<<(std::ostream& os, const Node& node);
-public:
-    Node(int i, int j, int g = 0, int h = 0, std::tuple<int, int, int> parentTuple = std::tuple<int, int, int>(-1, -1, -1));
-    std::tuple<int, int, int> getTuple() const;
-    std::tuple<int, int, int> getParentTuple() const;
-
-public:
-    int _i;
-    int _j;
-    int _g;
-    int _h;
-    int _time;
-    int _f;
-
-private:
-    BaseNode _node;
-    std::tuple<int, int, int> _parentTuple;
-};
-
-bool operator==(Node const& first, Node const& second);
-bool operator<(Node const& first, Node const& second);
-std::ostream& operator<<(std::ostream& os, Node const& node);
-
-struct hashNode {
-    typedef Node argument_type;
-    typedef std::size_t result_type;
-
-    result_type operator()(argument_type const& s) const;
-};
-
-struct hashTuple3 {
-    typedef std::tuple<int, int, int> argument_type;
-    typedef std::size_t result_type;
-
-    result_type operator()(argument_type const& s) const;
-};
 
 
 // SEARCH TREE
@@ -143,12 +77,3 @@ struct hashAgent {
 
     result_type operator()(argument_type const& s) const;
 };
-
-
-// std::ostream& operator<<(std::ostream& os, std::vector<Node> const& vec) {
-//     for (auto& v : vec) {
-//         os << v << "; ";
-//     }
-
-//     return os;
-// }
