+++ results/res0.txt
@@ -1,29 +1,29 @@
-0 0 0 0
-1 0 1 3
-2 0 2 4
-3 0 3 5
-3 1 4 4
-4 1 5 5
-5 1 6 6
-6 1 7 7
-6 2 8 6
-6 3 9 7
-6 4 10 8
-6 5 11 9
-5 5 12 8
-5 6 13 9
-4 6 14 8
-4 7 15 9
-4 8 16 10
-3 8 17 9
-2 8 18 8
-1 8 19 7
-0 8 20 6
-0 7 21 5
-0 6 22 4
-1 6 23 5
-1 5 24 4
-1 4 25 3
-0 4 26 2
-0 3 27 1
-0 2 28 0
+0 0
+1 0
+2 0
+3 0
+3 1
+4 1
+5 1
+6 1
+6 2
+6 3
+6 4
+6 5
+5 5
+5 6
+4 6
+4 7
+4 8
+3 8
+2 8
+1 8
+0 8
+0 7
+0 6
+1 6
+1 5
+1 4
+0 4
+0 3
+0 2
